infra/producer/producer.py
# Import Requirements
import time
import json
import os
import requests
from kafka import KafkaProducer
from dotenv import load_dotenv

load_dotenv()

# Define Variables for API
API_KEY = os.getenv("FINNHUB_API_KEY")
BASE_URL = os.getenv("FINNHUB_BASE_URL")
SYMBOLS = ["AAPL", "MSFT", "TSLA", "GOOGL", "AMZN"]

# Initialize Kafka Producer
producer = KafkaProducer(
    bootstrap_servers="localhost:29092",
    value_serializer=lambda v: json.dumps(v).encode("utf-8"),
)

# Retrieving the data
def fetch_quote(symbol):
    url = f"{BASE_URL}?symbol={symbol}&token={API_KEY}"
    try:
        response = requests.get(url)
        response.raise_for_status()
        data = response.json()
        data["symbol"] = symbol
        data['fetched_at'] = int(time.time())
        return data
    except Exception as e:
        print({f"Error fetching {symbol}: {e}"})
        return None
    
# Looping and Pushing to Stream
while True:
    for symbol in SYMBOLS:
        quote = fetch_quote(symbol)
        if quote:
            print(f"Producing: {quote}")
            producer.send("stock-quotes", value=quote)
    time.sleep(6)






